+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT5YV59PNZZ3EKRXRV416GDK"
+        "01KT66991ZPDKR90V2995SW8NB"
     )
     r"""The unique identifier of the data source"""