+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT5YV59KY6P4GZBYFYCN8AS6"
+        "01KT66991WWWV9SKD7F0ZFSVD9"
     )
     r"""The unique identifier of the data source"""