+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT5YV59R5K1JA0664D23HX6W"
+        "01KT6699228X0G4A038XBYZXMA"
     )
     r"""The unique identifier of the data source"""