+++ packages/orq-rc/RELEASES.md
@@ -8419,3 +8419,13 @@ Based on:
 - [python v4.10.0-rc.74] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.74] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.74 - packages/orq-rc
+
+## 2026-06-03 07:32:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.770.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.75] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.75] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.75 - packages/orq-rc